chore: split the massInPeaks function to a new file
